+++ extensions/Worker.Extensions.Tables/release_notes.md
@@ -4,6 +4,6 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.Tables 1.4.3
+### Microsoft.Azure.Functions.Worker.Extensions.Tables 1.5.0
 
-- Update `Microsoft.Azure.WebJobs.Extensions.Tables` to `1.3.3`.
+- Support deferred binding for poco types in Table Input binding (#3003)